extract test util methods
