+++ api/test/integration/api.test.js
@@ -4,37 +4,13 @@ const singleViewDb = require('../../lib/PostgresDb');
 const { loadSQL } = require('../../../api/lib/Utils');
 const { truncateTablesSQL } = loadSQL(path.join(__dirname, 'sql'));
 
-const BASE_URL = 'http://localhost:3010';
+const {
+  doDeleteRequest,
+  doPostRequest,
+  doGetRequest
+} = require('./TestUtils');
 
 describe('Singleview API', () => {
-  const rp = require('request-promise');
-  const doSearchRequest = async uri => {
-    const options = {
-      uri: `${BASE_URL}/${uri}`,
-      qs: {},
-      json: true
-    };
-    return await rp(options);
-  };
-
-  const doPostRequest = async (uri, body) => {
-    const options = {
-      method: 'POST',
-      uri: `${BASE_URL}/${uri}`,
-      body,
-      json: true
-    };
-    return await rp(options);
-  };
-
-  const doDeleteRequest = async uri => {
-    const options = {
-      method: 'DELETE',
-      uri: `${BASE_URL}/${uri}`
-    };
-    return await rp(options);
-  };
-
   beforeEach(() => {
     return singleViewDb.any(truncateTablesSQL);
   });
@@ -42,7 +18,7 @@ describe('Singleview API', () => {
   afterAll(singleViewDb.$pool.end);
 
   it('returns empty records for non-existent customer', async () => {
-    const response = await doSearchRequest(
+    const response = await doGetRequest(
       `customers?firstName=john&lastName=smith`
     );
     expect(response).toStrictEqual({
@@ -53,7 +29,7 @@ describe('Singleview API', () => {
   });
 
   it('returns uht record if customer exists in uht', async () => {
-    const response = await doSearchRequest(
+    const response = await doGetRequest(
       `customers?firstName=dani&lastName=beyn`
     );
     expect(response).toStrictEqual({
@@ -97,7 +73,7 @@ describe('Singleview API', () => {
     };
     await doPostRequest(`customers`, data);
 
-    const result = await doSearchRequest(
+    const result = await doGetRequest(
       `customers?firstName=Henrieta&lastName=sterre`
     );
 
@@ -133,7 +109,7 @@ describe('Singleview API', () => {
 
     expect(deleteResponse).toEqual('OK');
 
-    const result = await doSearchRequest(
+    const result = await doGetRequest(
       `customers?firstName=Henrieta&lastName=sterre`
     );
 
@@ -147,7 +123,7 @@ describe('Singleview API', () => {
   });
 
   it('returns academy-benefits record if customer exists in academy-benefits', async () => {
-    const response = await doSearchRequest(
+    const response = await doGetRequest(
       `customers?firstName=flor&lastName=beden`
     );
     expect(response).toStrictEqual({
@@ -172,7 +148,7 @@ describe('Singleview API', () => {
   });
 
   it('returns academy-councilTax record if customer exists in academy-councilTax', async () => {
-    const response = await doSearchRequest(
+    const response = await doGetRequest(
       `customers?firstName=alf&lastName=roscrigg`
     );
     expect(response).toStrictEqual({
@@ -197,7 +173,7 @@ describe('Singleview API', () => {
   });
 
   it('returns UHT-housingRegister record if customer exists in UHT-housingRegister', async () => {
-    const response = await doSearchRequest(
+    const response = await doGetRequest(
       `customers?firstName=Imojean&lastName=D'Abbot-Doyle`
     );
     expect(response).toStrictEqual({
@@ -222,7 +198,7 @@ describe('Singleview API', () => {
   });
 
   it('returns UHW record if customer exists in UHW', async () => {
-    const response = await doSearchRequest(
+    const response = await doGetRequest(
       `customers?firstName=Melisa&lastName=hansbury`
     );
     expect(response).toStrictEqual({
@@ -247,7 +223,7 @@ describe('Singleview API', () => {
   });
 
   it('Can get a record from Jigsaw', async () => {
-    const response = await doSearchRequest(
+    const response = await doGetRequest(
       `customers?firstName=firstname&lastName=lastname`
     );
     expect(response).toStrictEqual({
@@ -268,7 +244,7 @@ describe('Singleview API', () => {
   });
 
   it('Groups search records from all systems if they have same name, nino and dob', async () => {
-    const response = await doSearchRequest(
+    const response = await doGetRequest(
       `customers?firstName=elwira&lastName=moncur`
     );
     expect(response).toStrictEqual({
@@ -357,7 +333,7 @@ describe('Singleview API', () => {
   });
 
   it('Groups search records based on claim ID', async () => {
-    const response = await doSearchRequest(
+    const response = await doGetRequest(
       `customers?firstName=tate&lastName=Bullimore`
     );
     expect(response).toStrictEqual({
@@ -397,7 +373,7 @@ describe('Singleview API', () => {
   });
 
   it('Does not group search records only based on name', async () => {
-    const response = await doSearchRequest(
+    const response = await doGetRequest(
       `customers?firstName=Hartwell&lastName=lorinez`
     );
     expect(response).toStrictEqual({
